# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NFL Dead Money \u2014 Pundit Prediction Ledger API",
    "description": "\n## NFL Dead Money \u2014 Pundit Prediction Ledger API\n\nCryptographically-verified NFL analytics platform.  All prediction records are stored\nin an append-only BigQuery ledger with a tamper-evident SHA-256 hash chain.\n\n### Vendor Payloads\n\n| Payload | Endpoints | Description |\n|---|---|---|\n| **Pundit Index** | `GET /v1/leaderboard`, `GET /v1/pundits/` | Brier-scored accuracy rankings for tracked NFL media personalities |\n| **FMV Trajectory** | `GET /v1/cap/fmv/{player_name}` | Year-over-year Fair Market Value direction signal (improving / declining / flat) |\n| **Injury Lag** | `GET /v1/cap/players/{player_name}` | `availability_rating` \u00d7 `ml_risk_score` overlay identifying contracts not yet repriced for injury-adjusted market value |\n\n### Authentication\n\nB2B endpoints (`/v1/cap/*`) require an `X-API-Key` header.\nContact us to obtain a key.  Keys not set in `B2B_API_KEYS` env var \u2192 dev mode (auth disabled).\n\n### Rate Limits\n\nDefault: **1 000 requests / hour** per API key (configurable via `B2B_RATE_LIMIT_RPH`).\nExceeded quota returns HTTP 429 with a `Retry-After` header.\n",
    "contact": {
      "name": "Cap Alpha Protocol",
      "url": "https://github.com/ucalegon206/cap-alpha-protocol"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/cap/players": {
      "get": {
        "tags": [
          "b2b-cap-intelligence"
        ],
        "summary": "Paginated list of players with cap and FMV data",
        "description": "Returns paginated player cap data from the Gold layer.\nSorted by cap_hit_millions descending (highest earners first).",
        "operationId": "list_players_v1_cap_players_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by season year",
              "title": "Year"
            },
            "description": "Filter by season year"
          },
          {
            "name": "position",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by position (e.g. QB)",
              "title": "Position"
            },
            "description": "Filter by position (e.g. QB)"
          },
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by team abbreviation",
              "title": "Team"
            },
            "description": "Filter by team abbreviation"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapPlayersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cap/players/{player_name}": {
      "get": {
        "tags": [
          "b2b-cap-intelligence"
        ],
        "summary": "Single player cap profile \u2014 Injury Lag vendor payload",
        "description": "Returns the full cap profile for a player \u2014 all available years and all\nfinancial columns from fact_player_efficiency.",
        "operationId": "player_cap_profile_v1_cap_players__player_name__get",
        "parameters": [
          {
            "name": "player_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Player Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapPlayerProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cap/teams": {
      "get": {
        "tags": [
          "b2b-cap-intelligence"
        ],
        "summary": "Per-team cap summary",
        "description": "Returns pre-computed per-team cap aggregations from team_finance_summary.\nIncludes total cap committed, positional spending breakdowns, cap space,\nrisk cap (dead money), win totals, and conference.\nSorted by cap_space descending (most room first).",
        "operationId": "team_cap_summary_v1_cap_teams_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by season year",
              "title": "Year"
            },
            "description": "Filter by season year"
          },
          {
            "name": "conference",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "'AFC' or 'NFC'",
              "title": "Conference"
            },
            "description": "'AFC' or 'NFC'"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamCapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cap/fmv/{player_name}": {
      "get": {
        "tags": [
          "b2b-cap-intelligence"
        ],
        "summary": "Fair Market Value trajectory \u2014 FMV Trajectory vendor payload",
        "description": "Returns the year-over-year Fair Market Value (FMV) trajectory for a player.\nIncludes cap hit, FMV score, EDCE risk, efficiency ratio, and ML risk score.\n\nThe 'Pundit Index' vendor payload surfaces this endpoint to identify players\nwhose market compensation is diverging from on-field production value.",
        "operationId": "player_fmv_trajectory_v1_cap_fmv__player_name__get",
        "parameters": [
          {
            "name": "player_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Player Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FmvTrajectoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CapPlayerProfileResponse": {
        "properties": {
          "player_name": {
            "type": "string",
            "title": "Player Name",
            "example": "Patrick Mahomes"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/CapPlayerRecord"
            },
            "type": "array",
            "title": "Seasons",
            "description": "All available seasons, sorted by year descending"
          },
          "season_count": {
            "type": "integer",
            "title": "Season Count",
            "description": "Number of seasons on record",
            "example": 7
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "seasons",
          "season_count"
        ],
        "title": "CapPlayerProfileResponse",
        "description": "Response for GET /v1/cap/players/{player_name} \u2014 **Injury Lag** vendor payload.\n\nCross-reference `availability_rating` with `ml_risk_score` across seasons\nto detect contracts where availability decline has not yet been priced in."
      },
      "CapPlayerRecord": {
        "properties": {
          "player_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Name"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team",
            "example": "KAN"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year",
            "example": 2024
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position",
            "example": "QB"
          },
          "cap_hit_millions": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cap Hit Millions",
            "example": 45.0
          },
          "dead_cap_millions": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dead Cap Millions",
            "example": 10.0
          },
          "signing_bonus_millions": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Bonus Millions",
            "example": 20.0
          },
          "guaranteed_money_millions": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guaranteed Money Millions",
            "example": 210.0
          },
          "fair_market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fair Market Value",
            "example": 48.0
          },
          "ml_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ml Risk Score",
            "description": "XGBoost dead-cap risk probability [0, 1]. **Injury Lag** payload: high risk_score + low availability_rating signals a contract that has not repriced for injury exposure.",
            "example": 0.15
          },
          "edce_risk": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edce Risk",
            "example": 5.2
          },
          "availability_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Rating",
            "description": "games_played / max_games [0, 1]. **Injury Lag** payload: values below 0.75 with high cap_hit indicate an availability-repricing lag opportunity.",
            "example": 1.0
          },
          "games_played": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Games Played",
            "example": 17
          }
        },
        "type": "object",
        "title": "CapPlayerRecord",
        "description": "Single-season cap record for a player.\nThe `availability_rating` and `ml_risk_score` fields together form the\n**Injury Lag** vendor payload \u2014 identifying players whose contract has not\nyet repriced for declining injury-adjusted market value."
      },
      "CapPlayersResponse": {
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/CapPlayerRecord"
            },
            "type": "array",
            "title": "Players"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "players",
          "page",
          "limit",
          "total"
        ],
        "title": "CapPlayersResponse",
        "description": "Response for GET /v1/cap/players \u2014 paginated player cap data."
      },
...
      "FmvTrajectoryResponse": {
        "properties": {
          "player_name": {
            "type": "string",
            "title": "Player Name",
            "example": "Patrick Mahomes"
          },
          "trajectory": {
            "type": "string",
            "enum": [
              "improving",
              "declining",
              "flat",
              "unknown"
            ],
            "title": "Trajectory",
            "description": "Year-over-year FMV direction. **FMV Trajectory** vendor payload primary signal.",
            "example": "improving"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/FmvSeasonRecord"
            },
            "type": "array",
            "title": "Seasons",
            "description": "Per-season FMV records sorted by year ascending"
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "trajectory",
          "seasons"
        ],
        "title": "FmvTrajectoryResponse",
        "description": "Response for GET /v1/cap/fmv/{player_name} \u2014 the **FMV Trajectory** vendor payload.\n\nThe `trajectory` field is the primary signal for the vendor integration:\n- `improving` \u2014 FMV increased year-over-year (player value rising faster than cap hit)\n- `declining` \u2014 FMV decreased year-over-year (player value falling; contract at risk)\n- `flat`      \u2014 No meaningful change\n- `unknown`   \u2014 Insufficient history (fewer than 2 seasons)"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TeamCapRecord": {
        "properties": {
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team",
            "example": "KAN"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year",
            "example": 2024
          },
          "total_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cap",
            "description": "Total cap committed (millions USD)",
            "example": 230.0
          },
          "cap_space": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cap Space",
            "description": "Remaining cap space (millions USD)",
            "example": 25.4
          },
          "risk_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Cap",
            "description": "Dead money / at-risk cap (millions USD)",
            "example": 30.0
          },
          "qb_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qb Spending"
          },
          "wr_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wr Spending"
          },
          "rb_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rb Spending"
          },
          "te_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Te Spending"
          },
          "dl_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dl Spending"
          },
          "lb_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lb Spending"
          },
          "db_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Db Spending"
          },
          "ol_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ol Spending"
          },
          "k_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "K Spending"
          },
          "p_spending": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Spending"
          },
          "win_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Pct",
            "example": 0.824
          },
          "win_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Total",
            "example": 14.0
          },
          "conference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conference",
            "example": "AFC"
          }
        },
        "type": "object",
        "title": "TeamCapRecord"
      },
      "TeamCapResponse": {
        "properties": {
          "teams": {
            "items": {
              "$ref": "#/components/schemas/TeamCapRecord"
            },
            "type": "array",
            "title": "Teams"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "teams",
          "total"
        ],
        "title": "TeamCapResponse",
        "description": "Response for GET /v1/cap/teams"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}